move getUrl_ function for subscription offers into shared file
